refactor(file structure): move feature-status-bar/* to features/status-bar/*
